fix: Prompt for frozen account information located in oauth login

+++ api/services/account_service.py
@@ -302,7 +302,7 @@ class AccountService:
             from controllers.console.auth.error import EmailCodeAccountDeletionRateLimitExceededError
 
             raise EmailCodeAccountDeletionRateLimitExceededError(
-                cls.email_code_account_deletion_rate_limiter.time_window / 60
+                int(cls.email_code_account_deletion_rate_limiter.time_window / 60)
             )
 
         send_account_deletion_verification_code.delay(to=email, code=code)
@@ -451,7 +451,7 @@ class AccountService:
         if cls.reset_password_rate_limiter.is_rate_limited(account_email):
             from controllers.console.auth.error import PasswordResetRateLimitExceededError
 
-            raise PasswordResetRateLimitExceededError(cls.reset_password_rate_limiter.time_window / 60)
+            raise PasswordResetRateLimitExceededError(int(cls.reset_password_rate_limiter.time_window / 60))
 
         code, token = cls.generate_reset_password_token(account_email, account)
 
@@ -484,7 +484,7 @@ class AccountService:
         if cls.email_register_rate_limiter.is_rate_limited(account_email):
             from controllers.console.auth.error import EmailRegisterRateLimitExceededError
 
-            raise EmailRegisterRateLimitExceededError(cls.email_register_rate_limiter.time_window / 60)
+            raise EmailRegisterRateLimitExceededError(int(cls.email_register_rate_limiter.time_window / 60))
 
         code, token = cls.generate_email_register_token(account_email)
 
@@ -521,7 +521,7 @@ class AccountService:
         if cls.change_email_rate_limiter.is_rate_limited(account_email):
             from controllers.console.auth.error import EmailChangeRateLimitExceededError
 
-            raise EmailChangeRateLimitExceededError(cls.change_email_rate_limiter.time_window / 60)
+            raise EmailChangeRateLimitExceededError(int(cls.change_email_rate_limiter.time_window / 60))
 
         code, token = cls.generate_change_email_token(account_email, account, old_email=old_email)
 
@@ -565,7 +565,7 @@ class AccountService:
         if cls.owner_transfer_rate_limiter.is_rate_limited(account_email):
             from controllers.console.auth.error import OwnerTransferRateLimitExceededError
 
-            raise OwnerTransferRateLimitExceededError(cls.owner_transfer_rate_limiter.time_window / 60)
+            raise OwnerTransferRateLimitExceededError(int(cls.owner_transfer_rate_limiter.time_window / 60))
 
         code, token = cls.generate_owner_transfer_token(account_email, account)
         workspace_name = workspace_name or ""
@@ -727,7 +727,7 @@ class AccountService:
         if cls.email_code_login_rate_limiter.is_rate_limited(email):
             from controllers.console.auth.error import EmailCodeLoginRateLimitExceededError
 
-            raise EmailCodeLoginRateLimitExceededError(cls.email_code_login_rate_limiter.time_window / 60)
+            raise EmailCodeLoginRateLimitExceededError(int(cls.email_code_login_rate_limiter.time_window / 60))
 
         code = "".join([str(secrets.randbelow(exclusive_upper_bound=10)) for _ in range(6)])
         token = TokenManager.generate_token(
